+++ src/wav.c
@@ -114,8 +114,8 @@ static void write_wav(const char *fname, int16_t *samples, uint32_t num_samples,
 void write_mixdown_to_wav(char *filepath)
 {
     uint32_t len_sframes = proj->tl->out_mark_sframes - proj->tl->in_mark_sframes;
-    double *samples_L = get_mixdown_chunk(proj->tl, 0, len_sframes, true);
-    double *samples_R = get_mixdown_chunk(proj->tl, 1, len_sframes, true);
+    float *samples_L = get_mixdown_chunk(proj->tl, 0, len_sframes, true);
+    float *samples_R = get_mixdown_chunk(proj->tl, 1, len_sframes, true);
 
     int16_t *samples = malloc(sizeof(int16_t) * len_sframes);
     for (uint32_t i=0; i<len_sframes; i+=2) {
@@ -195,8 +195,8 @@ void load_wav_to_track(Track *track, const char *filename) {
     int16_t *src_buf = (int16_t *)wav_cvt.buf;
 
     for (uint32_t i=0; i<buf_len_samples; i+=2) {
-        clip->L[i] = (double) src_buf[i] / INT16_MAX;
-        clip->R[i] = (double) src_buf[i+1] / INT16_MAX;
+        clip->L[i] = (float) src_buf[i] / INT16_MAX;
+        clip->R[i] = (float) src_buf[i+1] / INT16_MAX;
     }
     memcpy(clip->L, wav_cvt.buf, audio_len_bytes);
     memcpy(clip->R, wav_cvt.buf, audio_len_bytes);
